# src/spa/src/dto/models/library-asset-for-list-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface LibraryAssetForListDto {
  assetType?: null | string;
  authorName?: null | string;
  copiesAvailable?: number;
  id?: number;
  title?: null | string;
  year?: number;
}