# src/spa/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Library Management System",
    "version": "V1"
  },
  "paths": {
...
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UserForLoginDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserForLoginDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserForLoginDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserForLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Author/{authorId}": {
      "get": {
        "tags": [
          "Author"
        ],
        "operationId": "GetAuthor",
        "parameters": [
          {
            "name": "authorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Author"
        ],
        "parameters": [
          {
            "name": "authorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Catalog": {
      "post": {
        "tags": [
          "Catalog"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryAssetForCreationDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryAssetForCreationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryAssetForCreationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryAssetForCreationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Catalog"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryAssetForUpdateDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryAssetForUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryAssetForUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryAssetForUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SearchString",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForListDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Catalog/{assetId}": {
      "delete": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Catalog"
        ],
        "operationId": "GetLibraryAsset",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Catalog/author/{authorId}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "authorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SearchString",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorDto": {
        "required": [
          "fullName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 250,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CategoryDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 250,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LibraryAssetAuthorDto": {
        "required": [
          "authorId",
          "librayAssetId",
          "order"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "librayAssetId": {
            "type": "integer",
            "format": "int32"
          },
          "authorId": {
            "type": "integer",
            "format": "int32"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "LibraryAssetCategoryDto": {
        "required": [
          "categoryId"
        ],
        "type": "object",
        "properties": {
          "librayAssetId": {
            "type": "integer",
            "format": "int32"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "LibraryAssetForCreationDto": {
        "required": [
          "assetAuthors",
          "assetCategories",
          "title",
          "year"
        ],
        "type": "object",
        "properties": {
          "deweyIndex": {
            "type": "string",
            "nullable": true
          },
          "assetType": {
            "$ref": "#/components/schemas/LibraryAssetTypeDto"
          },
          "assetAuthors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibraryAssetAuthorDto"
            }
          },
          "assetCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibraryAssetCategoryDto"
            }
          },
          "numberOfCopies": {
            "minimum": 0,
            "exclusiveMinimum": true,
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "maxLength": 250,
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "isbn": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "year": {
            "minLength": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "LibraryAssetForDetailedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "added": {
            "type": "string",
            "format": "date-time"
          },
          "numberOfCopies": {
            "type": "integer",
            "format": "int32"
          },
          "copiesAvailable": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "assetType": {
            "type": "string",
            "nullable": true
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorDto"
            },
            "nullable": true
          },
          "isbn": {
            "type": "string",
            "nullable": true
          },
          "deweyIndex": {
            "type": "string",
            "nullable": true
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LibraryAssetForListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "assetType": {
            "type": "string",
            "nullable": true
          },
          "authorName": {
            "type": "string",
            "nullable": true
          },
          "copiesAvailable": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "LibraryAssetForUpdateDto": {
        "required": [
          "assetAuthors",
          "assetCategories",
          "assetType",
          "copiesAvailable",
          "description",
          "id",
          "numberOfCopies",
          "title",
          "year"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "year": {
            "minLength": 1,
            "type": "integer",
            "format": "int32"
          },
          "numberOfCopies": {
            "minimum": 0,
            "exclusiveMinimum": true,
            "type": "integer",
            "format": "int32"
          },
          "copiesAvailable": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "maxLength": 250,
            "minLength": 1,
            "type": "string"
          },
          "assetType": {
            "$ref": "#/components/schemas/LibraryAssetTypeDto"
          },
          "isbn": {
            "type": "string",
            "nullable": true
          },
          "deweyIndex": {
            "type": "string",
            "nullable": true
          },
          "assetAuthors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibraryAssetAuthorDto"
            }
          },
          "assetCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibraryAssetCategoryDto"
            }
          }
        },
        "additionalProperties": false
      },
...
      "LibraryAssetTypeDto": {
        "enum": [
          "Book",
          "Media",
          "Other"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "ResetPassword": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UserForLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}